Oss release 2020-01-23-17-15
Changes:
- client and renderer creation rework
- three state visibility
- triangle strip and fan support
- pickable objects
- add android demo app
- add more dcsm support and examples
- remove mingw support
- remove static library installation
- add texture swizzling
- animation system removal
- c++ modernization

Original commit sha: 1c1e4fddf761507f4672c556caa7892436baab44

+++ external/capu/include/ramses-capu/util/Traits.h
@@ -124,11 +124,11 @@ namespace ramses_capu
         enum id
         {
             Identifier =
-                ramses_capu::is_CAPU_PRIMITIVE<T>::Value
-                | ramses_capu::is_CAPU_POINTER<T>::Value
-                | ramses_capu::is_CAPU_REFERENCE<T>::Value
-                | ramses_capu::is_CAPU_ENUM<T>::Value
-                | ramses_capu::is_CAPU_CLASS<T>::Value
+                static_cast<int>(ramses_capu::is_CAPU_PRIMITIVE<T>::Value)
+                | static_cast<int>(ramses_capu::is_CAPU_POINTER<T>::Value)
+                | static_cast<int>(ramses_capu::is_CAPU_REFERENCE<T>::Value)
+                | static_cast<int>(ramses_capu::is_CAPU_ENUM<T>::Value)
+                | static_cast<int>(ramses_capu::is_CAPU_CLASS<T>::Value)
         };
     };
